Added declarations, scopes, and type lookups.
Semicolon is no longer a valid token.

The initial indentation level is -1, so the whole source file is
now in a single block.

The soft separator token is now " separate", and it is removed from the
end if any.

Added class Declaration as a base for anything declarable.

Added class Type for any types.

Added typedef TypeSpec to represent composite types as a list of Type-s.

Added struct Match for the result of lookups, to be able to return
a list of matching type parameters of parametric types.

The pivot type is now the first item of the merged left-right arguments,
this is a temporary hack, and will be cleaned up.

Tuple labels are now represented as parametric types created on the fly.

The handling of the :function statement is now special cased.

Exceptions are rethrown to produce a core dump.