Generated code can use multiple registers by allocating them.
Added Regs struct to manage available register sets.

MOV to the same register is not encoded.

Added Storage::register field to store the actual register used.

Added Storage::swap bool field so that operand swapping don't have to
be signaled in the integer value field.

Added Regs argument to Value::compile methods.

Modified exponentiation_by_squaring to use RAX, RDX, and a value on the
stack, so that the same spilling code as the divmod can handle this case.

Rewrote integer arithmetic code generation REGISTER cases to use
an allocated register instead of the hardwired RAX. Added code to
handle REGISTER first operands, too (used to be forced onto the stack).

Added operation specific register allocation code.

Allow multiply, divmod, and exponent operations to leak the address
of a left MEMORY operand, so it can be used for the assignment version
of these operations, as they always need to be computed in registers.

Added spill and fill methods so that even when a certain operation
runs out of available registers, it can save some, use them, and
restore them at the end while putting the result on the stack.