Added boolean if, and fixed a bunch of bugs.
Modified Makefile because the core file is in the root directory.

Added explicit manipulations of BitSetOp enums.

Added BOGUS_TS so that some TypeSpec are explicitly meaningless.

Added rvalue and lvalue functions to add or remove the lvalue_type prefix.

Removed Value::set_ts, instead the constructor expects this value.

Added BooleanIf and BooleanIfValue. The formes is not yet used as statements
are handled in typize, not looked up yet.

Added SpeciaType class for types that can appear in the code, and may be
compiled to void, so that the Storage of these can be stored to NOWHERE
without the compiler complaining.

Added missing superclass call in FunctionBodyScope::allocate.

Added helper method Value::compile_to_void so the Storage returned by
compile does not get lost accidentally.

The right argument of integer assignments were wrongly expected to be an lvalue.

Added extra variable declarations, assignments, and if statement
to test source.