Memory management operations are put in separate functions.
Added X64::init_memory_management to set the native memory management
function labels, and generate the local ones.

One version of incref/decref is generated for each register.

Just one version is generated for alloc, using RAX for input/output,
and does not clobber any other registers anymore.

Register names are stored in an array.

A Label can ask for not getting an automatic def_index, which is handy if
it is supposed to be overwritten, but until then shouldn't be referenced.

The IMUL3 opcodes are now explicitly so, and can generate optimized
instruction with imm8 operand if the value allows.