Implemented generic and/or logical operators.
The return value is no longer boolean, but the pivot type,
which is the left type for or, and the right type for and.

Added Type::where method to query the optimal storage class.

Added Type::boolval method to compute the boolean value of a value
without destroying it.

Type::create methods can now create REGISTER things, too.

The arguments of the logical and operator are swapped in tupleize,
because the right argument is now the more important.

When checking the value of boolean registers and addresses, TESTB
instructions were replaced with CMPB, because it will be more
analogous with the more complex types.