Added ReferenceType with reference counting behavior.
Added HeapType to be the base of everything referrable. The
array_type is now a HeapType (and quite a dummy).

Variables declared to some HeapType are created with a reference
instead.

String literals are also returned as reference to an array.

Fixed logical operators accidentally forcing their arguments to bool.

Types with one parameter can now be looked up by name.

The creation of {alloc,free}_function moved to main.