Implemented lexical scopes for Code subexpressions.
Scope returned to the previous idea of not including itself in the
previous_declaration chain, because that would be ambiguous even
for starting the unwinding. Instead it uses Marker structs to
properly record the scope state to start finalizations.

Added Scop::intrude method to insert another Scope that takes
the later declarations.

Added declaration_cast helper function, because apparently
reinterpret_cast-ing Scope to Declaration goes crazy.

Fixed FunctionScope accidentally letting the body scope to
finalize the arguments and the result.

Rewrote Type methods that took a TypeSpecIter reference, because
altering the argument means it's impossible to do more than one
operation on it, which is necessary when trying multiple conversions.

BasicType now takes a parameter_count, too, because code_type
is a BasicType.

BasicType converts to code_type of itself, although it's a bit awkward.

Added helper constructor to TypeSpec to initialize from a TypeSpecIter.

The prints C function incorrectly used the width instead of the precision
of the %s directive to limit the characters.

The C code does a sanity check by counting the allocations.

The boolean if converts its arguments to CodeValue to put the branch into
a lexical scope.