String interpolation and empty parentheses fix.
SpecialType::where is now NOWHERE, so that Void functions can
return this storage properly.

Allow literal precedence stuff to follow each other, this is necessary
for the interpolating syntax.

When closing a group, the fore precedence is now LITERAL, because
if it's lower, then it could act on the following literal, and that
would be weird.

Empty groups are allowed at some places, so we have to deal with the
meaningless empty parentheses in tupleize. Silently ignore it
when it's the whole argument tuple, but complain in the individual
positional and keyword and pivot argument positions. Also
disallow labeled nothings.

The sanity check at the beginning of tupleize was wrong since the
zero index means no child node.

Simplified the cases when a pivot position labeling or separating
operator creates an implicit tuple.

Added handling of INITIALIZER typed Node-s and Expr-s.

Factored out the identifier lookup from typize, since it is now
used from a second place, too.

Interpret initializer expression acting on a string constant as
string interpolation. Encountered identifiers are looked up
without arguments, and have their stringify method called, and
the fragments concatenated.

Added brace_split function to util.cpp for splitting the template string.

Allow ArrayConcatenationValue receiving the right operand in the constructor
to simplify the concatenation of fragments.

FunctionValue now properly returns the return value.

StringValue stores the literal in std::string, and only converts to
characters when compiling.