Operator renamings.
The operators that can be both be used as unary and binary have names
prefixed with binary_ and unary_.

Assignment modifier operators have names prefixed with assign_.

Comparison operators are now named as is_ and not_.

Lookup of is_equal and not_equal falls back to looking up equality.

Boolean negation returns FLAGS storage, which is likely to be more optimal.