String interpolation implemented with streaming.
Added Function::get_pivot_typespec method.

typematch allows Code values to be converted.

Reordered functions in test.c a bit.

Added append_decode_utf8 C function to immediately append the decoded
characters to a character array.

Added append_other C function to just append characters to a character array.

Added streamify_ functions that take a character array lvalue, and
serialize their first argument to the end of it. Imported them in init_builtins.

Added TSs and Ss typedefs to init_builtins.

Added interpolate function to typize.cpp, which assembles operations to
serialize the desired value. A Character Array variable is declarad, initialized
to a funny string, then a series of streamify method calls is done one the
template fragments, then the variable value is returned. These are put in a
BlockValue for serial execution, and then in a CodeValue to enclose the
variable in a CodeScope, and clean it up when leaving the scope.

Added FunctionValue::force_arg to add a function argument directly.

Added FunctionValue::{push,pop}_arg helper methods for argument processing,
including some very incomplete support for passing lvalues as a
raw address on the stack.

Added BlockValue::force_add method to add statements directly.

Fixed CodeValue::compile not to return a MEMORY storage result, as it can
point to a variable to be destroyed.

Added DeclarationValue::force_variable method to declare a variable directly.

Added typespec.cpp to Makefile.