Simplified CodeScope usage.
CodeScope-s are no longer inserted anywhere, we know in advance
if it'll be needed or not.

Removed Scope::intrude method.

CodeValue now just gets the CodeScope that contains the subexpression's
declarations.

Marker-s are no longer added to all Value objects, only to
FunctionReturnValue, again.

String interpolation scopes itself like a statement would.

Added code_scoped_typize wrapper to help enclosing a subexpression
in a lexical scope.

BlockValue uses the Expr type to tell in advance if a statement
is a declaration, or not.