Moved built-in operations to type inner scopes.
Added inner scope to BooleanType and HeapType.

Added CharacterType class, with example zero and unicode initializers.

Added unicodeCharacterValue class for the latter.

Logical operations, and reference operations won't be scoped.

Inner scope is looked up for the meaningful type only, ignoring
attributes and reference specifications in TypeSpec-s.

Makefile uses zsh and pipefail option so that a failing compilation
does not proceed to testing.