Code scoping is again done with markers.
Added Value::marker variables, set by typize either when the type context
is Code, or for pivot values.

typematch converts values to CodeValue-s when the target TypeSpec is
Code.

CodeValue uses the marker in the Value to intrude its CodeScope.

The Boolean :if uses explicit calls to make_code_value instead of
calling typematch, but this is just a shortcut.

Removed code_scoped_typize.

FunctionReturnValue uses a Dummy Declaration to initiate finalization,
so it is immune to scoping changes.