Started using Pure type contexts.
The module body is typized with Pure context.

In typize the set_token calls are now done right after creating the
Value objects, so the check methods can report user errors properly.

BlockValue gets the type context. If that's Pure, all statements
are typized with Pure context. Otherwise all statements but the last one
are typized as Void Code, and the last one with the original context.

BlockValue::add_statement got a flag argument if to set the type
of the block itself.

Impure declarations in DeclarationValue and impure statements in
BlockValue now cause an error. For this, DeclarationValue sets its
type to Pure if the declaration was pure.

Temporarily removed typematch logging.

Updated Makefile with the new core_pattern.