Methods in records.
Prettified Storage printing with negative MEMORY offsets.

Added sanity check not to allocate a Variable twice.

Added self_scope to FunctionScope.

RecordType no longer allocate-s the inner scope, becaus that is already
added to another scope, and it will be invoked from there.

The $ is parsed as a single character token.

The $ is treeized as an identifier.

Added Node::NONE enum value, because the treeize loop variable kept its value,
and unless cleared, it contained the previous Node type, and invalid operators
looked like something valid.

BlockValue converts the statements to Void before butting it in a CodeValue,
to make sure that doesn't freak out when returning a MEMORY result value.

FunctionDefinitionValue adds a $ variable to the self scope if the definition's
scope has a type.

Record self is now passed by value.