Records passed by alias, including return values.
Fixed Variable not reserving the correct amount of bytes if an alias.

The internal function return variable no longer has an lvalue type.

The function self variable was never made an alias.

Added hidden function result variable to function calls. This variable's
address is passed to the called function in the result alias, and
the same address is used as the result of the function. Since this
variable is created in the enclosing scope, it must be destroyed by the
end of the statement.

The function return can use the return alias to store the result
value at that address.