Nameless record initializers.
RecordType only collects the member variables, and uses them to get
any further information about them.

Added RecordType::set_inner_scope method, because we need the member
variable list before the allocation phase. So removed the allocate method.

Added RecordInitializerValue class to perform the initialization
from a tuple.

The nameless initializer was renamed to "{}" in lookups for simplicity.

Added check_argument function to perform the conversion of a tuple
to a list of Value-s. Used by the RecordInitializerValue so far.