Declaration can avoid copying a newly constructed record.
Added Value::declare_dirty method to return an already declared
and initialized variable.

If a function call's pivot argument is a newly declared variable, and the
function is Void, so it would return the same variable, then
FunctionCallValue::declare_dirty returns that variable.

DeclarationValue can utilize the declare_dirty method of values, and
in that case it returns, but does not initialize (again) that variable.

Since it is the first declaration in a statement, escaping it means
escaping the first declaration, as implemented recently.