Transient regions in scopes don't overlap nontransient ones.
Each CodeScope now allocates first the nontransient items, then the
transient ones overlapping with each other. This is necessary to prevent
a transient CodeScope overlapping with the following nontransient
declaration, which would make it impossible to copy a transient variable
into the nontransient one. This happens when declaring a record variable
in a statement.