Multiple function return values.
Updated Function, FunctionDefinitionValue, FunctionCallValue, and
FunctionReturnValue to be capable of handling more than one value.

Added for far dummy multi_type, which is now the return type of
functions with more than one value.

Factored out FunctionCallValue::ret_res method to return a single
value, used by both returning one value, or returning the pivot value
in Void functions. The former means that returning a record alias
is now in a MEMORY with a dynamic variable, not an RBP based one.