Multiple return values unpacking.
Added MultiValue class for a rogue tuple, that doesn't become a block.
It needs a sequence of lvalues, and stores them all as ALISTACK.

Added UnpackingValue class that takes a MultiValue and a FunctionCallValue
with multiple return values, and performs the assignments from the
deconstructed return value to the multi components.

Added unpacking as assignment to Multi Lvalue.

typize only creates a BlockValue from a TUPLE if either the scope is
pure, or a type context was given. Otherwise it creates a MultiValue.

Function now typizes its body with a Void type context.

Added values/multi.cpp.