Generalized unpacking, works with explicitly packed values, too.
Added Value::unpack method to check if it is possible to unpack
a Value, and if yes, then retrieve the types.

MultiValue can decide itself if it's an lvalue or not. Rvalues
are stored on the stack as rvalue arguments.

Unpacking can create an implicit MultiValue from its positional
arguments, and does not expect a FunctionCallValue.