Unwinding is now the task of each Value object.
Added Value::unwind method to perform the unwinding when necessary.

jump_to_content_finalization is now a Scope method, because it is called
from the main unwinding function, and it takes a Declaration * instead
of a Marker.

The exception check in CodeScope was moved to CodeValue and
FunctionDefinitionValue.

Commented out old Unwind classes.

UnwindStack now stores Value pointers, and invokes their unwind
methods during the unwinding process.

Added stacked function to convert a StorageWhere to its stack target.