Added :switch and :when controls.
Added SwitchValue class to handle the outer block that stores the
switch expression's value into a hidden variable, and catches the
UNSWITCH exception.

Added WhenValue class to evaluate a method call of its expression
with the name of cover, and argument of the switch variable, and if
it's true, then execute the body and throw an UNSWITCH exception.

Added cover operation to integers, equivalent of the comparison.