Basic exception handling.
The exception symbol is now global, and referred from test.c.

Function stores the exception type it may raise for FunctionCallValue
to check.

FunctionScope also stores it for :raise to check.

Treenumeration metatype allows an initializer expression to declare
a concrete treenumeration type, although this mechanism is not yet
complete.

Built-in exceptions use negative numeric codes.

Added RaiseValue class to implement the :raise control.

FunctionDefinitionValue allows the use of a may argument that should
contain a treenum initializer for now.

FunctionCallValue checks for exceptions raised, and starts unwinding if necessary.