Added :for loop.
Added sanity check for CodeScope to make sure its contents are already
finalized when the enclosing scopen calls finalize on it.

Added ForEachValue class to implement the :for loop over an
iterator looking thing, which is now anything with a next method.

Added extra check to DeclarationValue to only allow declaration by
type context if the type is an lvalue.

Function ELF labels are now locals, except for the start function, so
multiple names can coexist for now.

Fixed FunctionCallValue::unwind not unwinding correctly if a pivot
value was used.

Fixed IntegerOperationValue not handling unary operations correctly,
expecting an extra argument.

Added explicit uncore Makefile target.