Using 64-bit absolute relocations in data, and Elf64 addends.
Ork now uses Elf64_Rela structures with an explicit addend.
This is necessary, because ld rewrites absolute relocations to
create dynamic relocations, and that uses explicit addends, so
implicit addends in the section are overwritten.

X64 passes the addends explicitly to Ork.

Relocations in the data section are now 64-bit absolute ones.

Updated enum stringifications and virtual method calls for this.

Lame attempt in enum and treenum to generate unique symbol names.