Streamifiable is now an interface implemented by various types.
Types implement Streamifiable by having an ImplementationType
inserted into their inner_scope.

Type::get_inner_scope gets a TypeSpecIter argument so it can be
used the same way as other Type methods. Sometimes a dummy iterator
is passed.

ImportedFunction collects every instance in a pointer vector, and
imports all of them by the static import_all method. This is necessary
as imported implementation methods are not accessible easily from
the root scope.

ImplementationType no longer needs the implementor type, as this type
will no longer be used as a ts of a Value, and won't participate in the
conversion process.

Type::autoconv is now implemented in the Type class, only get_inner_scope
method needs to be overridden.

Attribute and reference types forward the get_inner_scope method to
their type parameter.

MetaType forwards the allocate method to its inner_scope, as it's not
added into any other scope.

The rolematch function invokes the TypeSpec/Type autoconv methods.

Role matching is performed before checking references, so that a
conversion to an interface can match a reference, too. This may not be
necessary anymore.

Allow conversion of a reference to a nonreference, this is also
necessary for conversion to boolean, which was rejected until now.

Lookup of inner scopes are done by Value::lookup_inner method. This does
the usual thing by default, but the ImplementationConversionValue is
smart enough to retype itself to the original type, and do the lookup
in the implementation scope.

Refactored lookup function, so no separate lookup_scope function is used.
This is helpful as the lookup of the inner scopes is now implemented
in Value, so the argument check is now performed after the lookups.

Interpolation uses conversion into the streamifiable interface to look up
the streamify method in that implementation only.