The Iterator interface now implementes the Iterable interface.
This is kinda esoteric, but theoretically correct, even if the
code is a bit unclear at places.

Interfaces are now allowed to contain implementations of other
interfaces in their inner scope.

InterfaceType remembers its inner scope, because it must be found
by the recursive implementation lookup.

Replaces implementation_cast function with implemented. This function
checks for an implementation of the requested interface, and for
a match it does the same as Type::autoconv did, and wraps the Value
in an ImplementationConversionValue. If a non-matching implementation
it proceeds to the corresponding interface recursively.

Added IteratorIterableIterValue class to implement the Iterable iter
method by Iterator, which in fact just returns itself.

All these makes sense.