Item iteration works.
Added SameType class to allow Same typed variables to be measured
by returning a nonconcrete size of SAME_SIZE. Scopes sum these up
normally, then substitute the concrete size later. The where
of Same typed things is only defined for non-argument lvalues as MEMORY,
which is the same for all types.

Added create, store, and destroy methods to Variable, so that
it can first substitute type parameters to the declared Some type
before calling the similar methods on TypeSpec.

Scope::get_size now takes a TypeSpecIter argument to allow the
computation of the concrete size.