Using NO_TS instead of VOID_TS for not expecting any argument.
Renamed BOGUS_TS to NO_TS. It mustn't be the type on any Value,
only be used temporarily, or used to expect a NULL Value argument.

Expecting a Void pivot argument is forbidden, as it makes no sense.

Scope pivot hints default to NO_TS, and cannot be set to VOID_TS.

typematch doesn't accept NO_TS as the target type, this case must be
handled ouside of it by checking for a NULL Value.

typematch accepts VOID_TS as the target type, meaning any non-NULL Value,
whose return value is ignored, and unlike Void Code, it is not evaluated
in a separate scope. It is currently used for the :repeat init argument.
Void now matches Void.

Removed the <datatype> hack, and the autogenerated record equality
function using Expr-s.

ArgInfo::context now must be non-NULL always.