Inner scopes are now managed by their Type object.
The implement function gets the list of declarations to include in the
Implementation. Previously it finished the Implementation too early
before the items were added to the inner scope, which accidentally caused
no problem so far.

Renamed implement_* functions to define_* to make it less confusing.

Inner scopes are now created by calling a Type::make_inner_scope, and
it is also added to the outer scope of the Type object. It also requires
passing the inner scope's pivot type so it won't be forgotten.

Replaced Type::set_inner_scope with complete_type, practically doing the same.