Records also use explicit Initializer-s instead of plain methods.
String null initializer is implemented by the NullStringValue class
instead of a fake method.

Added RecordType::get_member_variables method.

Added PartialType::create method since record creation on the stack needs it.

Fixed PartialDeclarationvalue not calling partial->precompile.

We now differentiate between the Class and the Record initialization by
checking the Type following partial_type.

Initialization of records uses a TypeSpec { partial_type, lvalue_type, record_type },
which is kinda ugly, but works well, and is the simplest way to make sure
the storage of it will be ALIAS. But it won't be lvalue on the outside
to keep things simple.