Implemented Stack push and Queue {push,unshift} with autogrow.
Added *AutogrowValue classes to implement the growing of low level
containers wher the are full.

Added autogrow flag to ClassWrapperIdentifier, so it can insert an autogrow
operation between unwrapping the inner container, and performing the same
named operation on it.

The above operations are now implemented using the autogrowing unwrapping.

Removed obsoleted Stack/Queue classes for push and unshift.

Renamed StackInitializerValue to ClassWrapperInitializerValue, because it
is not specific to Stack, and is now also used by Queue.

Circularray unfolding/stretching is now part of its once compiled grow function.

Removed emptied values/stack.cpp.