Some typematch improvements.
Since we don't want Multi typed pivot values, typematch can scalarize
a value to be converted without worrying about undoing it on failure,
since such failure means a Multi couldn't be converted as a plain argument,
which is fatal. This allows further conversions after scalarization.

Added related check to lookup_unchecked to reject Multi pivots.

Swapped the multi and reference matching code, so a scalarized reference can be
matched correctly.

The source reference type is not dropped if the target is not a reference,
not sure why it was this way.