Added Map {add,remove,index} operations.
All three duplicates a bit the Set operations, because the arguments
these functions take are not the element type of the underlying Rbtree,
so some tweaking is necessary to allow comparison to a key type, even if
it is an incomplete Item, knowing that only the index part will be
actually compared.

Added values/map.cpp.

Modified Rbtree helper functions a bit to be more useful for Map. The has
now returns the found/allocated index in RDI. Similarly, add returns the
added/found index in RDI, whil making sure its value portion is
uninitialized.