Added reference borrowing to array indexing.
Modified RecordWrapperValue to pass a modified Args to the wrapped
operation's check method, instead of directly setting its right child.
This is necessary, as the check method is expected to be called always,
for example for array indexing.

Added Borrow subclass of Declaration.

Modified String operations for the new RecordWrapper usage.