Added Dvalue type, and the :for control works again.
Added dvalue_type, which is used as the type of the new retro-arguments,
(currently only as the each argument of :for), as the only context allowed
for bare declarations, and as the type of such declarations.

The storage is the same as lvalue, so the current solution still passes
the address of a local variable into the control. This variable is not
initialized by the caller, but created and destroyed by the control
before/after evaluating the control's body.

Added RetroVariable subclass of Variable that is not finalized, and used
for the above declarations.

DeclarationValue no longer creates an artificial value for bare
declarations, instead the value member remains NULL.