Cleanups for matchers.
Exception renamed to the more generic match_unmatched_exception_type.

Added explicit RaisingDummy subclass of Declaration, marking it as transient.

Raiser and the other cases now uses this as their unwinding point.

Using the previously added TransparentTryScope, but in a cleaner way,
it only keeps transient declarations, and flushes other to its outer scope.

The :is control now puts the match scope into the then scope, and does
not messes around with moving declarations.

But it checks the raised exception type to make sure it's UNMATCHED.