Replaced type_type with {generic,value,identity}type_type.
The Type constructor expects the explicit type parameter list as
an argument, instead of just a parameter count.

Type also stores which one of the new type_type-s it is.

A Multi containing types is now a Multi Generictype.

Removed Option {is,as} pseudo members.

Outer function no longer peek into Type::parameter_count, because
it's not there anymore.

Function return values and container type parameters are required
to be value types.