Replaced the *type_type approach with a TypeType enum.
Readded type_type as the single typename type, and Type-s now use
a TypeType enum to specify their kind. This made Type::match
cleaner, and we don't have to pretend that Any meant anything.

Also, the builtin initialization order is not so strict.

And now all types match where a type is expected, and we can give
meaningful errors when the wrong kind is used, instead of silently
skipping it and complaining about operations not found.