Added scope types, merged declare_* methods.
Added ScopeType enum, also a Scope constructor argument.

Removed Scope::is_{pure,arg} methods.

DeclarationValue calls the unified Value::declare method, which returns
a Declaration, and in a CODE_SCOPE it is further checked for a variable
to be initialized later.

TypeValue stores the potential pivot type in the check method, so it can
use it later in declare.