Added :Base control for base roles.
Replaced Allocable::xxx_is_allocated with a consistent handling of the
where field not being NOWHERE.

Added BaseRole subclass of Role, which allocates its data over the class
header, and does not reinitialize the VT pointer.

Type::complete_type method now returns bool to allow error checking.

Added Class::base_role field so the base role can be allocated first.

Fixed class initializer lookup casting a function call to Ref even if it
was just a call to a role initializer, which caused a wrong decref and
a crash.

Type::inner_scope is no longer added to the outer scope, just the inner
scope refers to the outer one. This does not need Type to become a Scope,
and since direct lookup into the inners scope is not necessary, it works.