Modified rbtree removal to avoid moving values in node swapping.
While copying values may not be a big deal performance wise, it was
broken for not updating the age linkage, and was even worse for the
upcoming finalizer callbacks that the WeakValueMap will set up, which
will use the indexes to identify a weak reference value. So keeping a
non-removed value at the same index is a nice feature, even if the code
is now a bit more longer.