VariableValue no longer forces Lvalue TypeSpec.
Added Variable::get_typespec to let the Variable decide how to access
it. Ovalue and Rvalue arguments became Lvalue-s, bot everything else
remains as declared. This makes it possible to declare internal variables,
like $, switch, eval, and exception variables not to be Lvalue-s.

As a side effect Implementation types' second type parameter, the implementor
type is no longer an Lvalue. This was an accidental change, but looks like
an improvement, as Lvalue deep inside a TypeSpec looked wrong anyway.