Autogrow cleanups.
Renamed all wrapped member names to "wrapped".

ClassWrapperIdentifier no longer messes with inner scopes and variables,
just looks up "wrapped" in its pivot.

MapAddValue no longer generates its own autogrow code, which was faulty
anyway, as it did not raise exceptions on borrowed containers. Now it
wraps its pivot in an autogrow, which does all the relevant operations.
The payload1 of the FCB is changed from the Map object's address to the
address of the wrapped reference, but that still works.

Added extra error message when compiling an unchecked Raiser.