Added WeakSet type.
Added Zero type, a data type with no actual storage, used as the
value type of Map when implementing WeakSet.

Allow NO_TS in check_arguments, in which case no argument must be
given for that ArgInfo, it makes argument handling more flexible.

MapAddValue can deal with no value type parameter, including not expecting
and handling a value argument.

Added MapRemoveValue::postresult method to allow subclasses to decide
how to return the result.

Added WeakSetType, which is actually a WeakIndexMap with a Zero value
type parameter, since this is the asiest way to implement this.