Require initialization of variables in code blocks with assignment.
Disabled the declaration by value feature, declarations now require
either a type name, or nothing.

The result of a declaration has the type
of T <Uninitialized>, and these don't convert to Void to avoid not
initializing them immediately.

Added built-in (Any <Uninitialized>) = operator to initialize variables
instead of plain assignment. This is required on all code block variables.

Added VOID_UNINITIALIZED_TS, as a special case used for bare declarations.
This will be the declaration's type until the initialization's right hand
side is evaluated, and its type is used to fix the declaration, and actually
declare something.

The make_declaration_by_value function is now make_initialization_by_value,
and creates explicit DeclarationValue and CreateValue objects.

Allow an assignment following a declaration without treating the
assignment as unary operator.

Fixed control typization potentially passing a NO_TS context to lookup.

Allow literal numbers in integer-like context be such integers.

check_argument treats ANY_TS context as no context.

CodeBlockValue reports errors for the statement, instead of reporting at the
end of the block, which was useless.

CodeBlockValue peeks into CreateValue initializations for declarations to
be escaped to the outer context.

DeclarationValue return type is now T <Uninitialized>, and it does not perform
any initializations itself.

Fixed ControlValue::set_context_ts not handling NULL contexts.

Also, Value::set_context_ts is now a void method.

The :is control allows initializations in the first argument to not be treated
as a value. This mess will be cleaned up soon.

CreateValue return type is T Lvalue, and is the MEMORY storage of the variable.

CreateValue makes sure the declared variable is the last declaration, so
no silly things like raising dummies follow it. This is valid, since the
initialization will happen in the final step.

Fixed CreateValue not calling precompile on the right hand side.

Updated interpolation to use the modified code.