Attempt to implement the equality matcher in a saner way.
Removed the ?= initializer operator, replaced with =, which already
works with <Uninitialized>.

Readded the ~= operator resolving to "match other" with ASSIGNING
precedence, and processed as the equality matcher.

Fixed ControlValue::check_args not being contextless with Any type.

The :is control processes the positional argument as Any, but requires
that it raises UNMATCHED. Implicit equality matching does not work
at the moment.

Fixed :yield requiring a Void argument in Void context, instead of
requiring nothing.

Moved some comparison related Value classes from value.cpp to generic.cpp.

Reimplemented EqualityMatcherValue to do the comparison instead of
expecting that in its child Value.