Equality matching works implicitly, too.
Added EqualitymatcherType to be used as a type context in :is controls
within :switch, so initializers of the switch value type can create
the corresponding equality matchers.

Fixed BasicType::compare not working with CONSTANT storages.

Added ImplicitEqualityMatcherValue class to implement equality matching
to a non-initializer value.

Added InitializerEqualityMatcherValue class to implement equality
matching to an initializer value.

Factored out creation of a VariableValue to a switch variable into
lookup_switch function.

Hacked check_argument so that a context of <Equalitymatcher> does not
need a typematching on the resulting argument value.

The :is control provides a T <Equalitymatcher> context to its positional
argument, so initializers will create an equality matcher to the corresponding
initialized value.